Support CDN usage
Using Bootstrap from CDN requires an exact version number that is not
tied to gem releases.  Although historically `Bootstrap::VERSION` has
mirrored the official Bootstrap version number, this might not always be
the case.  Therefore, this patch adds `Bootstrap::VENDOR_VERSION`, which
is updated by the updater task.

Additionally, it is best practice to employ subresource integrity hashes
when using any public CDN.  To support this, this patch adds
`Bootstrap::VENDOR_INTEGRITY`, which is a Hash mapping asset names to
their properly-encoded integrity hashes.  This Hash is also updated by
the updater task.